+++ src/exceptions/RuntimeException.ts
@@ -1,0 +1,18 @@
+import Logger from "@bejibun/logger";
+import {defineValue} from "@bejibun/utils";
+
+export default class RuntimeException extends Error {
+    public code: number;
+
+    public constructor(message?: string, code?: number) {
+        super(message);
+        this.name = "RuntimeException";
+        this.code = defineValue(code, 500);
+
+        Logger.setContext(this.name).error(this.message).trace(this.stack);
+
+        if (Error.captureStackTrace) {
+            Error.captureStackTrace(this, RuntimeException);
+        }
+    }
+}