+++ src/exceptions/index.ts
@@ -1,3 +1,4 @@
+export * from "@/exceptions/ExceptionHandler";
 export * from "@/exceptions/ModelNotFoundException";
 export * from "@/exceptions/RouterInvalidException";
 export * from "@/exceptions/ValidatorException";